Better conf file reading
It checks all the listed files in the directory:
By that it fixes an issue that user can have one .znvmrc in the home
directory, it was impossible to override it by a local .nvmrc

So now it checks the directory for .znvmrc, .nvmrc and Dockerfile, and
if any of them found, it will be used.

/home/me/.znvmrc
/home/me/src/.nvmrc
/home/me/src/project/Dockerfile

Always the most specific will win, in the project folder the Dockerfile,
in any other folder inside the src folder the .nvmrc and in any other
folder the .znvmrc.