Support finding closest version from bottom to top
If the znvmrc contains full version, like 12.12.12 but only 12.1.1 the
plugin would say that the version is not found. However there must be
not be any breaking change between the two, we would still need to
install both of them.
Instead of requiring to install every single version, match closest
version: If you require 12.12.12, first we will try to find the exact
version, if it's not installed, try to find by minor version (12.12),
if it's still not found it will try to find by major version (12).