Improve documentation of basic setup
Using the `znvm current` command we can check if nodejs is configured in
the current PATH. If yes, we don't necessary need to call the `znvm use
default` command. It happens if we open a subshell when znvm was already
configured. For example we open a terminal, where `znvm use default`
runs, then we open a tmux. The PATH will be already filled with the
nodejs path, so we don't need to call it again.